+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a fn to poll DMA transfers
 - Add initial support for LEDC in ESP32-H2 (#560)
 - Add initial support for ASSIST_DEBUG in ESP32-H2 (#566)
+- Add all `SPI` examples for the ESP32-H2 (#549)
 
 ### Changed